First rough pass at new layout code. Addresses:  - Need to specify variable axis widths  - No padding between chart and first axis  - Removal of whitelisting of chart components
To reach these goals axes need to be specified within an AxisGroup. The Axis group may be left or right aligned. Axes within the group can specify a width. Similarly charts need to be specified within a ChartGroup. Anything within the ChartGroup is assumed to be a chart. It will be passed a scale and dimensions and is expected to return an SVG <g> containing the chart of other overlaid visualization. For example, two line charts where each line chart uses a different axis:

<ChartContainer beginTime={beginTime} endTime={endTime} padding="5">
	<ChartRow height="200">
		<AxisGroup align="left">
			<YAxis id="axis1" min={0} max={2} width="50" type="linear" format="$,.2f"/>
		</AxisGroup>
		<ChartGroup>
			<LineChart axis="axis1" data={audCurrency} classed="aud"/>
			<LineChart axis="axis2" data={euroCurrency} classed="euro"/>
		</ChartGroup>
		<AxisGroup align="right">
			<YAxis id="axis2" min={0.65} max={0.85} width="50" type="linear" format="$,.2f"/>
		</AxisGroup>
	</ChartRow>
</ChartContainer>